bug fix for diagonal right or left checking of winner.
